moved scrapes into worker container
- created new dockerfile for worker
- api dockerfile is now alpine based
- moved files into src folder (so that supervisor doesnt have to watch node_modules folder)
