+++ Dockerfile
@@ -1,7 +1,4 @@
-#switch to -slim when building in conatiner
-#currently linking to source folder since easier to dev...
-FROM node:8-slim
-RUN  apt-get update &&  apt-get install -y build-essential python
+FROM node:8-alpine
 WORKDIR /build
 COPY package.json /build
 RUN npm install --production
@@ -10,4 +7,4 @@ COPY . /build
 
 EXPOSE 8080
 USER node
-CMD ["node","index.js"]
+CMD ["node","src/api/index.js"]
