+++ src/api/index.js
@@ -6,221 +6,35 @@ const hapiCron = require('hapi-cron');
 const moment = require('moment');
 const boom = require('boom');
 const url = require('url');
-const classifier = require('ammobin-classifier');
-const throat = require('throat');
+const RedisSMQ = require('rsmq');
+const helpers = require('../helpers');
 
-const axios = require('axios');
-const version = require('./package.json').version;
-const axiosVersion = require('./node_modules/axios/package.json').version;
-axios.defaults.headers.common['User-Agent'] = `AmmoBin.ca/${version} (nodejs; Linux x86_64) axios/${axiosVersion}`; // be a nice web citizen and tell people who we are..
-
-const client = redis.createClient({ host: 'redis' });
+const { SOURCES, DATE_FORMAT, CACHE_REFRESH_HOURS, QUEUE_NAME } = require('../constants');
 const influx = require('./influx');
 
-const cabelas = require('./cabelas-api');
-const canadiantire = require('./canadian-tire');
-const wolverinesupplies = require('./wolverine-api');
-const theAmmoSource = require('./the-ammo-source');
-const hirsch = require('./hirschprecision');
-const wildWest = require('./wild-west');
-const tigerArms = require('./tiger-arms');
-const magdump = require('./mapdump');
-const rangeviewsports = require('./rangeviewsports');
-const jobrook = require('./jo-brook');
-const faoc = require('./faoc');
-const alflahertys = require('./alflahertys');
-const bullseyelondon = require('./bullseyelondon');
-const sail = require('./sail');
-const reliablegun = require('./reliablegun');
-const tenda = require('./tenda');
-const canadaammo = require('./canadaammo');
-const frontierfirearms = require('./frontierfirearms');
-const tradex = require('./tradex');
-const bvoutdoors = require('./bvoutdoors');
+const rsmq = new RedisSMQ({ host: 'redis' });
+const client = redis.createClient({ host: 'redis' });
 
-const helpers = require('./helpers');
-
-const PROXY_URL = 'https://images.ammobin.ca';
-const DATE_FORMAT = 'YYYY-MM-DD';
-const CACHE_REFRESH_HOURS = 4;
 const secretRefreshKey = Math.random().toString(); // required to only allow us to refresh the cache on our terms
 console.log('secretRefreshKey', secretRefreshKey)
 
-function proxyImages(items) {
-  return items.map(i => {
-    if (!i.img) {
-      return i;
-    }
-
-    if (i.img.indexOf('//') === 0) {
-      i.img = 'http://' + i.img;
-    }
-    i.img = PROXY_URL + '/x160/' + i.img;
-    return i;
-  })
-}
-
-function addSrcRefToLinks(items) {
-  return items.map(i => {
-    if (!i.link) {
-      return i;
-    }
-
-    if (i.link.indexOf('?') === -1) {
-      i.link += '?'
-    } else {
-      i.link += '&'
-    }
-    i.link = `https://api.ammobin.ca/track-outbound-click?url=${encodeURIComponent(i.link + 'utm_source=ammobin.ca')}`;
-    return i;
-  });
-}
-
-function classifyBrand(items) {
-  return items.map(i => {
-    i.brand = classifier.classifyBrand(i.brand || i.name || '')
-    return i;
-  });
-}
-
-function getCounts(items) {
-  return items.map(i => {
-    i.count = isNaN(i.count) ? classifier.getItemCount(i.name) || '' : i.count;
-    if (i.count > 1) {
-      i.unitCost = i.price / i.count;
-      if (i.unitCost < 0.01) {
-        i.unitCost = null; // something went wrong with the classifier
-      }
-    } else {
-      i.unitCost = null;
-    }
-    return i;
-  });
-}
-
-function makeSearch(source, type) {
-  switch (source) {
-    case 'cabelas.ca':
-      return cabelas(type);
-
-    case 'firearmsoutletcanada.com':
-      return faoc(type);
-
-    case 'alflahertys.com':
-      return alflahertys(type);
-
-    case 'bullseyelondon.com':
-      return bullseyelondon(type);
-
-    case 'sail.ca':
-      return sail(type);
-
-    case 'reliablegun.com':
-      return reliablegun(type);
-
-    case 'canadiantire.ca':
-      return canadiantire(type);
-
-    case 'gotenda.com':
-      return tenda(type);
-
-    case 'canadaammo.com':
-      return canadaammo(type);
-
-    case 'jobrookoutdoors.com':
-      return jobrook(type);
-
-    case 'theammosource.com':
-      return theAmmoSource(type);
-
-    case 'hirschprecision.com':
-      return hirsch(type);
-
-    case 'gun-shop.ca':
-      return wildWest(type);
-
-    case 'tigerarms.ca':
-      return tigerArms(type);
-
-    case 'magdump.ca':
-      return magdump(type);
-
-    case 'rangeviewsports.ca':
-      return rangeviewsports(type);
-
-    case 'wolverinesupplies.com':
-      return wolverinesupplies(type);
-
-    case 'frontierfirearms.ca':
-      return frontierfirearms(type);
-
-    case 'tradeexcanada.com':
-      return tradex(type);
-
-    case 'bvoutdoors.com':
-      return bvoutdoors(type);
-
-    default:
-      throw new Error(`unknown source: ${source} + type: ${type}`);
+rsmq.listQueues(function (err, queues) {
+  if (err) {
+    console.error(err)
+    throw err;
   }
-}
-
-function getKey(source, type) {
-  return `${moment.utc().format(DATE_FORMAT)}_${source}_${type}`;
-}
-
-function getItems(source, type) {
-
-  console.log('making scrape for ', source, type, new Date());
-  return makeSearch(source, type)
-    .then(classifyBrand)
-    .then(i => proxyImages(i))
-    .then(i => addSrcRefToLinks(i))
-    .then(getCounts)
-    .then(items => items.filter(i => i.price))
-    .then(items => {
-
-      if (items.length) {
-        console.log(`found ${items.length} ${source} ${type}`);
-      } else {
-        console.warn(`WARN: no results for ${source} ${type}`, new Date())
+  if (queues.indexOf(QUEUE_NAME) === -1) {
+    rsmq.createQueue({ qname: QUEUE_NAME }, function (err, resp) {
+      if (err) {
+        console.error(err)
+        throw err;
       }
-
-      const key = getKey(source, type);
-      return new Promise((resolve, reject) => client.set(key, JSON.stringify(items), 'EX', 172800 /*seconds => 48hrs*/, (err) => {
-        if (err) {
-          return reject(err);
-        }
-        return resolve(items);
-      }));
-    })
-    .catch(e => {
-      console.error(`ERROR: failed to load ${source} ${type} => ${e && e.message ? e.message : e}`);
+      if (resp === 1) {
+        console.log("queue created")
+      }
     });
-}
-
-const SOURCES = [
-  'canadiantire.ca',
-  'sail.ca',
-  'alflahertys.com',
-  'firearmsoutletcanada.com',
-  'bullseyelondon.com',
-  'reliablegun.com',
-  'cabelas.ca',
-  'gotenda.com',
-  'canadaammo.com',
-  'wolverinesupplies.com',
-  'jobrookoutdoors.com',
-  'theammosource.com',
-  'hirschprecision.com',
-  'gun-shop.ca',
-  'tigerarms.ca',
-  'magdump.ca',
-  'rangeviewsports.ca',
-  'frontierfirearms.ca',
-  'tradeexcanada.com',
-  'bvoutdoors.com',
-];
+  }
+});
 
 
 const server = new Hapi.Server({
@@ -244,7 +58,7 @@ const classifiedListsCache = server.cache({
   segment: 'classifiedLists',
   generateFunc: function (type, next) {
     // pull everything out of cache
-    const keys = SOURCES.map(s => getKey(s, type));
+    const keys = SOURCES.map(s => helpers.getKey(s, type));
 
     return new Promise((resolve, reject) =>
       client.mget(keys, (err, res) => err ? reject(err) : resolve(res.map(r => r ? JSON.parse(r) : null)))
@@ -316,7 +130,7 @@ const bestPricesCache = server.cache({
   expiresIn: (CACHE_REFRESH_HOURS * 2) * 60 * 1000, // make sure that this can stay around till after the next refresh cycle
   segment: 'bestPrices',
   generateFunc: function (id, next) {
-    const keys = SOURCES.map(s => getKey(s, 'centerfire'));
+    const keys = SOURCES.map(s => helpers.getKey(s, 'centerfire'));
     return new Promise((resolve, reject) =>
       client.mget(keys, (err, res) => err ? reject(err) : resolve(res))
     )
@@ -398,7 +212,7 @@ server.route({
   method: 'GET',
   path: '/dank',
   handler: function (request, reply) {
-    const keys = SOURCES.map(s => getKey(s, 'centerfire'));
+    const keys = SOURCES.map(s => helpers.getKey(s, 'centerfire'));
     return new Promise((resolve, reject) => {
       client.mget(keys, (err, res) => err ? reject(err) : resolve(res.map(r => r ? JSON.parse(r) : null)))
     })
@@ -442,7 +256,7 @@ server.route({
   handler: function (req, reply) {
 
     const type = req.query.type;
-    console.log(req.query['secret-refresh-key'], secretRefreshKey)
+    console.log('secret-refresh-key', req.query['secret-refresh-key'], secretRefreshKey)
 
     if (req.query['secret-refresh-key'] !== secretRefreshKey) {
       return reply(boom.badRequest('fuck off. you dont have my super secret code...'))
@@ -450,11 +264,10 @@ server.route({
       return reply(boom.badRequest('invalid type: ' + type));
     }
 
-    const throttle = throat(1); // make it REALLY slow so we dont run out of memory
-    return Promise.all(SOURCES.map(s => throttle(() => getItems(s, type))))
-      .then(() => new Promise((resolve, reject) =>
-        classifiedListsCache.drop(type, (e, r) => e ? reject(e) : resolve(r)))
-      )
+    return Promise.all(SOURCES.map(source =>
+      new Promise((resolve, reject) => rsmq.sendMessage({ qname: QUEUE_NAME, message: JSON.stringify({ source, type }) },
+        (err, res) => err ? reject(err) : resolve(res)))
+    ))
       .then(() => new Promise((resolve, reject) =>
         bestPricesCache.drop(type, (e, r) => e ? reject(e) : resolve(r)))
       )
