+++ src/random/move-clicks.js
