+++ src/scrapes/hirschprecision.js
@@ -1,4 +1,4 @@
-const helpers = require('./helpers');
+const helpers = require('../helpers');
 
 
 module.exports = function (type) {
