--- test.js
@@ -1,3 +1,0 @@
-const ff = require('./frontierfirearms');
-
-ff('centerfire').then(r => console.log(r))
